add shop create edit

+++ database/migrations/2021_09_09_193406_create_order_product_table.php
@@ -15,8 +15,8 @@ class CreateOrderProductTable extends Migration
     {
         Schema::create('order_product', function (Blueprint $table) {
             $table->id();
-            $table->integer('order_id')->unsigned();
-            $table->integer('product_id')->unsigned();
+            $table->bigInteger('order_id')->unsigned();
+            $table->bigInteger('product_id')->unsigned();
             $table->integer('qty')->unsigned();
             $table->integer('unit_price')->unsigned()->nullable();
             $table->foreign('order_id')->references('id')->on('orders')->onDelete('no action');
